requirements.txt
asn1crypto==0.24.0
boto3==1.7.16
botocore==1.10.16
cffi==1.11.5
cryptography==2.2.2
docutils==0.14
enum34==1.1.6
futures==3.2.0
idna==2.6
ipaddress==1.0.22
jmespath==0.9.3
pycparser==2.18
pyOpenSSL==17.5.0
python-dateutil==2.7.2
s3transfer==0.1.13
scandir==1.7
six==1.11.0
